requirements.txt
langgraph==1.0.7
langchain-community==0.4.1
langchain-text-splitters==1.1.0
langchain-ollama==1.0.0
langchain-groq==1.1.2
chromadb==1.4.1
pillow==12.1.0
pymupdf==1.26.7
torch==2.10.0
ultralytics==8.4.10
pathlib==1.0.1
huggingface_hub==1.3.7
transformers==5.0.0
python-pptx==0.6.23
